requirements.txt
bs4
pyquery
datetime
elasticsearch
dateutil.parser
